rest_api_jwt_mini_project_final/requirements.txt
Flask~=2.0.3
PyJWT~=1.7.1
Werkzeug~=2.0.3